make simulator use `sql_generation` crate as dependency
