+++ simulator/Cargo.toml
@@ -16,15 +16,14 @@ path = "main.rs"
 
 [dependencies]
 turso_core = { path = "../core", features = ["simulator"]}
-rand = "0.8.5"
-rand_chacha = "0.3.1"
+rand = { workspace = true }
+rand_chacha = "0.9.0"
 log = "0.4.20"
 env_logger = "0.10.1"
 regex = "1.11.1"
 regex-syntax = { version = "0.8.5", default-features = false, features = [
     "unicode",
 ] }
-anarchist-readable-name-generator-lib = "=0.1.2"
 clap = { version = "4.5", features = ["derive"] }
 serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
@@ -32,9 +31,10 @@ notify = "8.0.0"
 rusqlite.workspace = true
 dirs = "6.0.0"
 chrono = { version = "0.4.40", features = ["serde"] }
-tracing = "0.1.41"
+tracing = { workspace = true }
 tracing-subscriber = { version = "0.3.19", features = ["env-filter"] }
 anyhow.workspace = true
-turso_sqlite3_parser = { workspace = true, features = ["serde"]}
 hex = "0.4.3"
 itertools = "0.14.0"
+sql_generation = { workspace = true }
+turso_parser = { workspace = true }
